backend/requirements.txt
# saas-core library from GitHub
git+https://github.com/Symbiot01/saas-core.git

# Core dependencies for saas-core library
pyjwt[crypto]>=2.8.0
cryptography>=41.0.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Optional: Database support (placeholder for future use)
sqlalchemy[asyncio]>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0

# Example dependencies (for backend examples)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.9

# S3/DigitalOcean Spaces connector
boto3>=1.28.0

# DICOM processing
pydicom>=2.4.0

# DICOM compression support
pylibjpeg>=2.0.0
pylibjpeg-libjpeg>=2.1.0

# LangChain for AI analysis
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
